qcacld-2.0: Fix the issue of WPA3 connection failure
Failed to connect to WPA3 AP with RSN IE containing multiple
AKMs, because CSR_RSN_MX_AUTH_SUITES is 4, if the position
of SAE (SHA256) is after the fourth, it cannot be found since
only the first four items are listed in the AuthSuites variable.

The maximum number of AKMs should be 59, to fix this issue define
CSR_RSN_MX_AUTH_SUITES as 59.

Change-Id: I1a25ec129b79c0c116f3c951bacf98b9eba48bbc
CRs-Fixed: 4217751